backend/requirements.txt
asgiref==3.8.0
#backports.zoneinfo==0.2.1
Django==4.2.11
django-cors-headers==4.3.1
django-perms==2.1.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pathlib==1.0.1
pathlib2==2.3.7.post1
pillow==10.3.0
pytz==2024.1
PyYAML==6.0.1
#six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-110/six-1.15.0-py2.py3-none-any.whl
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
whitenoise==6.6.0
gunicorn==22.0.0
django-heroku==0.3.1